refactor: separate event logic from calendar logic
